Allow deno_cli to be used as lib
